Refactor company detail sections
