--- src/components/CompanyDetail.jsx
@@ -1,8 +1,0 @@
-const handleAddAssignment = async (assignment) => {
-  try {
-    await addAssignment(assignment);
-    // fetchAssignments artık gerekli değil, addAssignment içinde otomatik çağrılıyor
-  } catch (error) {
-    console.error('Atama eklenirken hata oluştu:', error);
-  }
-};
